Move async file.exists test out of EditorManager into DocumentCommandHandlers, to keep EditorManager’s chowCustomViewer and it’s affiliates synchronous.
At the same time this fixes unit tests that were broken by merging pull request 5910 / to fix 5742

However the original facet of bug 5742, where an image file is deleted while being viewed and then clicked on in the file tree is still not fixed, because the expectation is not accurate. When a file is deleted it stays in view, if you click on its file in the project tree it remains in view.

This fixes the case where you have an image file in the project tree that has been deleted from disk but you are looking at another file, either text or image. If you then click on a stale image the file open error will pop up.